Address PR review: rename submit_eval_jobs scripts; add --olmo_eval_ref; deprecate old script. Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
